Took out active admin and devise. Made user registration to circumvent user show route issue. All tests pass.
