--- app/admin/user.rb
@@ -1,18 +1,0 @@
-ActiveAdmin.register User do
-  permit_params :first_name, :last_name, :linkedin, :twitter, :email
-
-  # See permitted parameters documentation:
-  # https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
-  #
-  # permit_params :list, :of, :attributes, :on, :model
-  #
-  # or
-  #
-  # permit_params do
-  #   permitted = [:permitted, :attributes]
-  #   permitted << :other if resource.something?
-  #   permitted
-  # end
-
-
-end
