+++ app/helpers/user_registrations_helper.rb
@@ -1,0 +1,2 @@
+module UserRegistrationsHelper
+end
