+++ app/views/user_registrations/new.html.haml
@@ -1,7 +1,7 @@
 - content_for(:page_title) do
-  %h1 "Put in yo stuff
+  %h1 "Put in yo stuff"
 
-= simple_form_for @user, url: user_path do |f|
+= simple_form_for @user, url: user_registration_path do |f|
   = f.input :first_name
   = f.input :last_name
   = f.input :email
