+++ app/views/users/index.html.haml
@@ -7,4 +7,4 @@
 
 %ul.user_pagination
   - @users.each do |user|
-    %li.user= link_to "#{user.last_name}, #{user.first_name}", users_path(user)
+    %li.user= link_to "#{user.last_name}, #{user.first_name}", user_path(user)
